backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Arpan Mondal",email = "arpanmondal572@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
